% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.THEMEcomp.R
\name{THEME}
\alias{THEME}
\title{THEME}
\usage{
THEME(Xlist, Xnew = NULL, E, nbcomp, s = 0.5, l = 1,
  OutputDir = NULL, cvvChoice = NA, bwopondChoice = NA,
  updateProgress = NULL, myEps = 10^(-6))
}
\arguments{
\item{Xlist}{List of R matrix thematic blocks of the thematic model used for calibration.}

\item{Xnew}{List of R matrix thematic blocks of the thematic model used for validation (the default-value NULL causes the prediction-error rates to be calculated on the calibration data).}

\item{E}{Design matrix assigning a role to each theme in each equation: rows correspond to the equations, columns to the themes, and each entry of E may be Y,X or A, according to whether the theme is dependent, explanatory, or additional-explanatory in the equation.
When a theme is declared additional-explanatory, no component will ba calculated in it, and its variables will enter the model directly as additional covariates.}

\item{nbcomp}{Vector containing the numbers of components to extract in themes.}

\item{s}{Parameter that tunes the balance of the structural relevance (SR) of components with respect to the goodness of fit (GOF) in the THEME criterion.
Value 1 means that only the SR will be taken into account, and not the GOF. Value 0 means that only the GOF will be taken into account, and not the SR. Default-value .5 gives equal importance to SR and GOF.}

\item{l}{Parameter that tunes the locality of variable-bundles the components should focus on. Default 1 is the minimum value and indicates that no local focussing is to be performed.}

\item{OutputDir}{Path Directory where to save the outputs (default "C:\\").}

\item{cvvChoice}{[lko] Number of units to be left out in cross-validation (default NA indicates no cross-validation is to be performed).}

\item{bwopondChoice}{[bwsw] Weight of SR in the backward-selection criterion (default NA indicates that no backward selection is to be performed).}

\item{updateProgress}{argument restricted to THEME.UI (keep the NULL default).}
}
\description{
THEmatic Model Exploration is a both exploratory and predictive multiblock-multiequation-technique.
The dataset is partitioned by the user into thematic blocks of variables. The thematic blocks are linked by a thematic model consisting of one or several equations.
Each equation links a dependent theme to explanatory themes. THEME searches each block for a given number of components capturing the theme's information that best serves the overall model.
Then, THEME can perform cross-validation-based backward component-selection to identify the number of really useful components in themes.
This backward selection produces a decreasing sequence of models, each of which is associated with a vector of component-numbers in themes.
THEME outputs the prediction-error rates of dependent variables for each model in the sequence, so that the user can compare the performance of the models in detail.
THEME also outputs the loadings of variables on components in all themes, the coefficients of the component-regression models, among other relevant numeric information.
}
\keyword{THEME}
